feat: InspectEvm fn renames, inspector docs, book cleanup
