+++ book/src/SUMMARY.md
@@ -2,8 +2,8 @@
 
 - [Introduction](./../../README.md)
 - [Awesome Revm](./awesome.md)
-- [Dev section](./dev.md)
 - [Architecture](./architecture.md)
+- [Dev section](./dev.md)
 - [Revme](./revme.md)
 - [Release procedure](./release_procedure.md)
 - [Contact](./contact.md)
