+++ book/src/architecture.md
@@ -1,4 +1,4 @@
-# Architecture
+# Architecture and API
 
 REVM is a flexible implementation of the Ethereum Virtual Machine (EVM). It follows the rules of the Ethereum mainnet and stays up to date with changes through hardforks as defined in the official [Ethereum 
 execution specs](https://github.com/ethereum/execution-specs).
@@ -7,26 +7,45 @@ You can use REVM in two main ways:
 1. Run regular Ethereum transactions using a Execution API
 2. Create your own custom version of the EVM (for Layer 2 solutions or other chains) using EVM framework
 
-The main `revm` library combines all the different crates into one package and reexports them. You can see overview revm crates in [crates folder](https://github.com/bluealloy/revm/tree/main/crates) and overview of examples in [examples folder](https://github.com/bluealloy/revm/tree/main/examples).
+To see usage examples you can check the [examples folder](https://github.com/bluealloy/revm/tree/ main/examples). Other than documentation, examples are main resource to see and learn about Revm.
 
-REVM works in no_std environments which means it can be used in zero-knowledge virtual machines (zkVMs). It also has very few external dependencies, which you can read more about in the [dev section](./dev.md).
+The main `revm` library combines all crates into one package and reexports them, standalone library are useful if there is need to import functionality with smaller scope. You can see overview of revm crates in [crates folder](https://github.com/bluealloy/revm/tree/main/crates).
+
+REVM works in `no_std` environments which means it can be used in zero-knowledge virtual machines (zkVMs) and it is the standard library in that use case. It also has very few external dependencies.
 
 # Execution API
 
-REVM provides four ways to execute transactions through traits (interfaces).
+`Evm` the main structure for executing mainnet ethereum transaction is build with a `Context` and a builder, code for it looks like this:
 
-The State system builds on the `Database` trait and handles:
-- Getting data from external storage
-- Managing the EVM's output
-- Caching changes when running multiple transactions
+```rust,ignore
+let mut evm = Context::mainnet().with_block(block).build_mainnet();
+let out = evm.transact(tx);
+```
 
-You can implement one of three database interfaces depending on what you need:
+`Evm` struct contains:
+* `Context` - Environment and evm state.
+* `Instructions` - EVM opcode implementations
+* `Precompiles` - Built-in contract implementations
+* `Inspector` - Used for tracing.
 
-- `Database`: Uses a mutable reference (`&mut self`). This is useful when you want to update a cache or track statistics while getting data. Enables basic transaction functions like `transact` and `inspect`.
+And `Context` contains data used in execution:
+* Environment data, the data that is known before execution starts are `Transaction`, `Block`, `Cfg`.
+* `Journal` is place where internal state is stored. Internal state is returned after execution ends.
+   * And `Database` is a interface that allows fetching external data that is needed in runtime. That data are account, storage and bytecode. When loaded they are stored in `Journal` 
 
-- `DatabaseRef`: Uses a regular reference (`&self`). Good for when you only need to read data without making changes. Enables reference-based functions like `transact_ref`.
+REVM provides four ways to execute transactions through traits (API):
 
-- `Database + DatabaseCommit`: Adds the ability to save transaction changes directly. Enables commit functions like `transact_commit`.
+* `transact(tx)` and `replay()` are function of `ExecuteEvm` trait that allow execution transactions. They return the status of execution with reason, changed state and in of case of failed execution a error.
+* `transact_commit(tx)` and `replay_commit()` are part of `ExecuteCommitEvm` that internally commits the state diff to the database and returns status of execution. Database is requires to support `DatabaseCommit` trait.
+* `inspect()`, `inspect_replay(tx)` and few others are part of `InspectEvm` trait that allow execution with inspection. This is how tracers are called.
+* `inspect_commit()`,`inspect_replay_commit(tx)` that are part of `InspectCommitEvm` trait that extends `InspectEvm` to allow commiting state diff after tracing.
+
+For inspection API to be enabled `Evm` needs to be create with inspector.
+
+```rust,ignore
+let mut evm = Context::mainnet().with_block(block).build_mainnet().with_inspector(inspector);
+let _ = evm.inspect_with_tx(tx);
+```
 
 # EVM Framework
 
@@ -38,16 +57,12 @@ Each trait needed to build custom EVM has detailed documentation explaining how 
 
 In summary, REVM is built around several key traits that enable customizable EVM functionality. The core traits include:
 
-* **EvmTr**: The core EVM trait that provides access to the main EVM components:
-  - Context - Environment and state access
-  - Instructions - EVM opcode implementations
-  - Precompiles - Built-in contract implementations
-  - Inspector - Used for tracing, only enabled with `InspectorEvmTr` trait.
-* **ContextTr**: Accessed through EvmTr, defines the execution environment including Tx/Block/Journal/Db:
+* **EvmTr**: The core EVM trait that provides access to `Context`, `Instruction`, `Precompiles`:
+* **ContextTr**: Accessed through EvmTr, defines the execution environment including Tx/Block/Journal/Db.
 * **Handler**: Implements the core execution logic, taking an EvmTr implementation. The default implementation follows Ethereum consensus.
 
 And traits that provide support for inspection and tracing:
 
-* **InspectorEvmTr**: Extends EvmTr to enable inspection mode execution with an associated Inspector type
-* **InspectorHandler**: Extends Handler with inspection-enabled execution paths that make Inspector callbacks
+* **InspectorEvmTr**: Extends EvmTr to enable inspection mode execution with an associated `Inspector` type
+* **InspectorHandler**: Extends Handler with inspection-enabled execution paths that make `Inspector` callbacks
 * **Inspector**: User-implementable trait for EVM inspection/tracing
