+++ book/src/awesome.md
@@ -17,6 +17,7 @@ A curated list of excellent Revm-related resources. Feel free to contribute to t
 - [**Reth:**](https://github.com/paradigmxyz/reth) A modular, contributor-friendly, and ultra-fast implementation of the Ethereum protocol.
 - [**Helios**](https://github.com/a16z/helios) is a trustless, efficient, and portable multichain light client written in Rust.
 - [**Trin**](https://github.com/ethereum/trin) is a Rust implementation of a Portal Network client.
+- 
 
 #### EVM Variants
 - [**Optimism**](https://github.com/bluealloy/revm/tree/main/crates/optimism) is a ethereum L2 network.
