--- book/src/testing.md
@@ -1,1 +1,0 @@
-# Testing
