+++ crates/inspector/src/lib.rs
@@ -1,4 +1,7 @@
-//! Optimism-specific constants, types, and helpers.
+//! Inspector is a crate that provides a set of traits that allow inspecting the EVM execution.
+//!
+//! It is used to implement tracers that can be used to inspect the EVM execution.
+//! Implementing inspection is optional and it does not effect the core execution.
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 #![cfg_attr(not(feature = "std"), no_std)]
 
@@ -8,6 +11,7 @@ extern crate alloc as std;
 #[cfg(all(feature = "std", feature = "serde-json"))]
 mod eip3155;
 mod gas;
+pub mod handler;
 mod inspect;
 mod inspector;
 mod mainnet_inspect;
@@ -21,6 +25,7 @@ pub mod inspectors {
     pub use super::gas::GasInspector;
 }
 
+pub use handler::{inspect_instructions, InspectorHandler};
 pub use inspect::{InspectCommitEvm, InspectEvm};
 pub use inspector::*;
 pub use noop::NoOpInspector;
