+++ crates/inspector/src/mainnet_inspect.rs
@@ -1,16 +1,20 @@
+use crate::{
+    handler::inspect_instructions,
+    inspect::{InspectCommitEvm, InspectEvm},
+    Inspector, InspectorEvmTr, InspectorFrame, InspectorHandler, JournalExt,
+};
 use context::{ContextSetters, ContextTr, Evm, JournalOutput, JournalTr};
 use database_interface::DatabaseCommit;
 use handler::{
     instructions::InstructionProvider, EthFrame, EvmTr, EvmTrError, Frame, FrameResult, Handler,
     MainnetHandler, PrecompileProvider,
 };
-use interpreter::{interpreter::EthInterpreter, FrameInput, InterpreterResult};
-
-use crate::{
-    inspect::{InspectCommitEvm, InspectEvm},
-    Inspector, InspectorEvmTr, InspectorFrame, InspectorHandler, JournalExt,
+use interpreter::{
+    interpreter::EthInterpreter, FrameInput, Interpreter, InterpreterAction, InterpreterResult,
+    InterpreterTypes,
 };
 
+// Implementing InspectorHandler for MainnetHandler.
 impl<EVM, ERROR, FRAME> InspectorHandler for MainnetHandler<EVM, ERROR, FRAME>
 where
     EVM: InspectorEvmTr<
@@ -24,6 +28,7 @@ where
     type IT = EthInterpreter;
 }
 
+// Implementing InspectEvm for Evm
 impl<CTX, INSP, INST, PRECOMPILES> InspectEvm for Evm<CTX, INSP, INST, PRECOMPILES>
 where
     CTX: ContextSetters + ContextTr<Journal: JournalTr<FinalOutput = JournalOutput> + JournalExt>,
@@ -46,6 +51,7 @@ where
     }
 }
 
+// Implementing InspectCommitEvm for Evm
 impl<CTX, INSP, INST, PRECOMPILES> InspectCommitEvm for Evm<CTX, INSP, INST, PRECOMPILES>
 where
     CTX: ContextSetters
@@ -54,10 +60,50 @@ where
     INST: InstructionProvider<Context = CTX, InterpreterTypes = EthInterpreter>,
     PRECOMPILES: PrecompileProvider<CTX, Output = InterpreterResult>,
 {
-    fn inspect_commit_previous(&mut self) -> Self::CommitOutput {
+    fn inspect_replay_commit(&mut self) -> Self::CommitOutput {
         self.inspect_replay().map(|r| {
             self.ctx().db().commit(r.state);
             r.result
         })
+    }
+}
+
+// Implementing InspectorEvmTr for Evm
+impl<CTX, INSP, I, P> InspectorEvmTr for Evm<CTX, INSP, I, P>
+where
+    CTX: ContextTr<Journal: JournalExt> + ContextSetters,
+    I: InstructionProvider<
+        Context = CTX,
+        InterpreterTypes: InterpreterTypes<Output = InterpreterAction>,
+    >,
+    INSP: Inspector<CTX, I::InterpreterTypes>,
+{
+    type Inspector = INSP;
+
+    fn inspector(&mut self) -> &mut Self::Inspector {
+        &mut self.data.inspector
+    }
+
+    fn ctx_inspector(&mut self) -> (&mut Self::Context, &mut Self::Inspector) {
+        (&mut self.data.ctx, &mut self.data.inspector)
+    }
+
+    fn run_inspect_interpreter(
+        &mut self,
+        interpreter: &mut Interpreter<
+            <Self::Instructions as InstructionProvider>::InterpreterTypes,
+        >,
+    ) -> <<Self::Instructions as InstructionProvider>::InterpreterTypes as InterpreterTypes>::Output
+    {
+        let context = &mut self.data.ctx;
+        let instructions = &mut self.instruction;
+        let inspector = &mut self.data.inspector;
+
+        inspect_instructions(
+            context,
+            interpreter,
+            inspector,
+            instructions.instruction_table(),
+        )
     }
 }
