+++ crates/optimism/src/api/exec.rs
@@ -106,7 +106,7 @@ where
     INSP: Inspector<CTX, EthInterpreter>,
     PRECOMPILE: PrecompileProvider<CTX, Output = InterpreterResult>,
 {
-    fn inspect_commit_previous(&mut self) -> Self::CommitOutput {
+    fn inspect_replay_commit(&mut self) -> Self::CommitOutput {
         self.inspect_replay().map(|r| {
             self.ctx().db().commit(r.state);
             r.result
